chore: moved nx structs to their own file
